[git/base_repository] (RK-16) extract #git method to base class
The BareRepository and WorkingRepository classes both need to run git
commands but the initial versions of these classes just copied the old
`#git` method. This commit extracts a base class and adds the #git
method to that class.
