+++ lib/r10k/git/bare_repository.rb
@@ -1,8 +1,9 @@
 require 'r10k/git'
+require 'r10k/git/base_repository'
 require 'r10k/logging'
 
 # Create and manage Git bare repositories.
-class R10K::Git::BareRepository
+class R10K::Git::BareRepository < R10K::Git::BaseRepository
 
   # @return [Pathname] The path to this Git repository
   def git_dir
@@ -61,51 +62,5 @@ class R10K::Git::BareRepository
     else
       :unknown
     end
-  end
-
-  include R10K::Logging
-
-  private
-
-  # Wrap git commands
-  #
-  # @param cmd [Array<String>] cmd The arguments for the git prompt
-  # @param opts [Hash] opts
-  #
-  # @option opts [String] :path
-  # @option opts [String] :git_dir
-  # @option opts [String] :work_tree
-  # @option opts [String] :raise_on_fail
-  #
-  # @raise [R10K::ExecutionFailure] If the executed command exited with a
-  #   nonzero exit code.
-  #
-  # @return [String] The git command output
-  def git(cmd, opts = {})
-    raise_on_fail = opts.fetch(:raise_on_fail, true)
-
-    argv = %w{git}
-
-    if opts[:path]
-      argv << "--git-dir"   << File.join(opts[:path], '.git')
-      argv << "--work-tree" << opts[:path]
-    else
-      if opts[:git_dir]
-        argv << "--git-dir" << opts[:git_dir]
-      end
-      if opts[:work_tree]
-        argv << "--work-tree" << opts[:work_tree]
-      end
-    end
-
-    argv.concat(cmd)
-
-    subproc = R10K::Util::Subprocess.new(argv)
-    subproc.raise_on_fail = raise_on_fail
-    subproc.logger = self.logger
-
-    result = subproc.execute
-
-    result
   end
 end
