+++ lib/r10k/git/working_repository.rb
@@ -1,9 +1,9 @@
 require 'r10k/git'
-require 'r10k/logging'
+require 'r10k/git/base_repository'
 require 'r10k/git/alternates'
 
 # Manage a non-bare Git repository
-class R10K::Git::WorkingRepository
+class R10K::Git::WorkingRepository < R10K::Git::BaseRepository
 
   # @attribute [r] path
   #   @return [Pathname]
@@ -65,51 +65,5 @@ class R10K::Git::WorkingRepository
   # @return [String] The origin remote URL
   def origin
     git(['config', '--get', 'remote.origin.url'], :path => @path.to_s).stdout
-  end
-
-  include R10K::Logging
-
-  private
-
-  # Wrap git commands
-  #
-  # @param cmd [Array<String>] cmd The arguments for the git prompt
-  # @param opts [Hash] opts
-  #
-  # @option opts [String] :path
-  # @option opts [String] :git_dir
-  # @option opts [String] :work_tree
-  # @option opts [String] :raise_on_fail
-  #
-  # @raise [R10K::ExecutionFailure] If the executed command exited with a
-  #   nonzero exit code.
-  #
-  # @return [String] The git command output
-  def git(cmd, opts = {})
-    raise_on_fail = opts.fetch(:raise_on_fail, true)
-
-    argv = %w{git}
-
-    if opts[:path]
-      argv << "--git-dir"   << File.join(opts[:path], '.git')
-      argv << "--work-tree" << opts[:path]
-    else
-      if opts[:git_dir]
-        argv << "--git-dir" << opts[:git_dir]
-      end
-      if opts[:work_tree]
-        argv << "--work-tree" << opts[:work_tree]
-      end
-    end
-
-    argv.concat(cmd)
-
-    subproc = R10K::Util::Subprocess.new(argv)
-    subproc.raise_on_fail = raise_on_fail
-    subproc.logger = self.logger
-
-    result = subproc.execute
-
-    result
   end
 end
